requirements.txt
fastapi
uvicorn
streamlit
requests
pandas
plotly
pytest
python-dotenv
alphagenome
